src/third_party/blink/web_tests/http/tests/priority-hints/script-async-initial-load.html

<script src=resources/common.js></script>
<script src=../resources/testharness.js></script>
<script src=../resources/testharnessreport.js></script>

<script id=script1 async importance=high src=../resources/dummy.js></script>
<script id=script2 async importance=low src=../resources/dummy.js?1></script>
<script id=script3 async importance=auto src=../resources/dummy.js?2></script>
<script id=script4 async importance=xyz src=../resources/dummy.js?3></script>
<script id=script5 async src=../resources/dummy.js?4></script>

<script>
  async_test(t => {
    addEventListener('DOMContentLoaded', t.step_func(() => {
      const msg = 'all scripts were fetched by the preload scanner';
      assert_true(internals.isPreloaded(script1.src), msg);
      assert_true(internals.isPreloaded(script2.src), msg);
      assert_true(internals.isPreloaded(script3.src), msg);
      assert_true(internals.isPreloaded(script4.src), msg);
      assert_true(internals.isPreloaded(script5.src), msg);

      assert_equals(getPriority(script1.src), kHigh,
                    'high importance on async <script> raises the priority ' +
                    'to kHigh');
      assert_equals(getPriority(script2.src), kLow,
                    'low importance on async <script> has no effect');
      assert_equals(getPriority(script3.src), kLow,
                    'auto importance on async <script> has no effect');
      assert_equals(getPriority(script4.src), kLow,
                    'invalid importance on async <script> has no effect');
      assert_equals(getPriority(script5.src), kLow,
                    'missing importance on async <script> has no effect');

      t.done();
    }));
  }, 'importance attribute on async <script>s fetched by the ' +
     'preload scanner takes the correct effect on resource load priority');
</script>
